add config runner
